+++ src/nativelib/core/__init__.py
@@ -1,20 +1,27 @@
-"""Library for read and write clickhouse native format."""
+"""Classes and functions for read and write Clickhouse blocks."""
 
-from .core import (
-    Array,
+from . import sizes as Size
+from .blocks import (
     BlockReader,
     BlockWriter,
-    ClickhouseDtype,
+)
+from .columns import (
     Column,
     ColumnInfo,
-    DType,
-    LowCardinality,
+)
+from .errors import (
     NativeLibError,
     NativeLibValueError,
-    Size,
 )
-from .reader import NativeReader
-from .writer import NativeWriter
+from .types.dtypes import (
+    DTypeFunc,
+    ClickhouseDtype,
+)
+from .types.objects import (
+    Array,
+    DType,
+    LowCardinality,
+)
 
 
 __all__ = (
@@ -25,12 +32,9 @@ __all__ = (
     "Column",
     "ColumnInfo",
     "DType",
+    "DTypeFunc",
     "LowCardinality",
-    "NativeReader",
     "NativeLibError",
     "NativeLibValueError",
-    "NativeWriter",
     "Size",
 )
-__author__ = "0xMihalich"
-__version__ = "0.2.4.dev0"